Added support for UserProfile Extractions and removed debug junk
Field names are now surrounded in quotes allowing for attributes with periods in them (listviews). Stopped going down the road of generating CSOM calls into the user profile service since you can just use csrShim with a ListView Web Part directly and get all the nice values that way. For CBQWPs the jsLink will just need to take care of this as needed. This makes the bulk of the fieldMappings work pretty worthless. I've left it in for now since it doesn't run unless a fieldmapping is passed and it may come in handy later. Probably not though. Sigh. Either way, objectives accompilshed!