Removed all FieldMapping templates
Removed the FieldMapping parameter and all the supporting functions (can still be found in the FieldMappings branch). This makes for a much smaller file and simpler approach. Probably doesn't affect performance too much, but that stuff was fairly redundant and went against the idea of matching the OOTB CTX generation as much as possible.